requirements.txt
Django>=2
requests>=2
graphene-django>=2
Pillow>=5
django-filter>=2
django-cors-headers>=2
psycopg2_binary>=2
celery>=4
django-graphql-jwt>=0.2
boto3>=1.9
django-storages>=1.7
